+++ java-manta-client-unshaded/src/main/java/com/joyent/manta/client/MantaClient.java
@@ -182,16 +182,6 @@ public class MantaClient implements AutoCloseable {
     private final UriSigner uriSigner;
 
     /**
-     * Object containing saved context used between requests to the Manta client.
-     */
-    private final MantaConnectionContext connectionContext;
-
-    /**
-     * Connection factory instance used for building requests to Manta.
-     */
-    private final MantaConnectionFactory connectionFactory;
-
-    /**
      * MBean supervisor.
      */
     private final MantaMBeanSupervisor beanSupervisor;
@@ -209,19 +199,16 @@ public class MantaClient implements AutoCloseable {
      * @param config The configuration context that provides all of the configuration values.
      */
     public MantaClient(final ConfigContext config) {
-        dumpConfig(config);
 
-        final String mantaURL = config.getMantaURL();
-        final String account = config.getMantaUser();
+        dumpConfig(config);
 
         ConfigContext.validate(config);
 
-        this.url = mantaURL;
         this.config = config;
-        this.home = ConfigContext.deriveHomeDirectoryFromUser(account);
+        this.url = config.getMantaURL();
+        this.home = ConfigContext.deriveHomeDirectoryFromUser(config.getMantaUser());
 
-        final KeyPairFactory keyPairFactory = new KeyPairFactory(config);
-        final KeyPair keyPair = keyPairFactory.createKeyPair();
+        final KeyPair keyPair = new KeyPairFactory(config).createKeyPair();
 
         final Signer.Builder builder = new Signer.Builder(keyPair);
         if (ObjectUtils.firstNonNull(
@@ -232,15 +219,13 @@ public class MantaClient implements AutoCloseable {
         final ThreadLocalSigner signer = new ThreadLocalSigner(builder);
         this.signerRef = new WeakReference<>(signer);
 
-        this.connectionFactory = new MantaConnectionFactory(config, keyPair, signer);
-        this.connectionContext = new MantaApacheHttpClientContext(this.connectionFactory);
+        final MantaConnectionFactory connectionFactory = new MantaConnectionFactory(config, keyPair, signer);
+        final MantaApacheHttpClientContext connectionContext = new MantaApacheHttpClientContext(connectionFactory);
 
         if (BooleanUtils.isTrue(config.isClientEncryptionEnabled())) {
-            this.httpHelper = new EncryptionHttpHelper(connectionContext, connectionFactory,
-                    config);
+            this.httpHelper = new EncryptionHttpHelper(connectionContext, config);
         } else {
-            this.httpHelper = new StandardHttpHelper(connectionContext, connectionFactory,
-                    config);
+            this.httpHelper = new StandardHttpHelper(connectionContext, config);
         }
 
         this.uriSigner = new UriSigner(this.config, keyPair, signer);
@@ -266,26 +251,21 @@ public class MantaClient implements AutoCloseable {
                 final ThreadLocalSigner signer) {
         dumpConfig(config);
 
-        final String mantaURL = config.getMantaURL();
-        final String account = config.getMantaUser();
-
         ConfigContext.validate(config);
 
-        this.url = mantaURL;
+        this.url = config.getMantaURL();
         this.config = config;
-        this.home = ConfigContext.deriveHomeDirectoryFromUser(account);
+        this.home = ConfigContext.deriveHomeDirectoryFromUser(config.getMantaUser());
 
         this.signerRef = new WeakReference<>(signer);
 
-        this.connectionFactory = connectionFactory;
-        this.connectionContext = new MantaApacheHttpClientContext(this.connectionFactory);
+        final MantaApacheHttpClientContext connectionContext =
+                new MantaApacheHttpClientContext(connectionFactory);
 
         if (BooleanUtils.isTrue(config.isClientEncryptionEnabled())) {
-            this.httpHelper = new EncryptionHttpHelper(connectionContext, connectionFactory,
-                    config);
+            this.httpHelper = new EncryptionHttpHelper(connectionContext, config);
         } else {
-            this.httpHelper = new StandardHttpHelper(connectionContext, connectionFactory,
-                    config);
+            this.httpHelper = new StandardHttpHelper(connectionContext, config);
         }
 
         this.uriSigner = new UriSigner(this.config, keyPair, signer);
@@ -440,7 +420,7 @@ public class MantaClient implements AutoCloseable {
         Validate.notBlank(rawPath, "rawPath must not be blank");
 
         final String path = formatPath(rawPath);
-        final HttpGet get = connectionFactory.get(path);
+        final HttpGet get = httpHelper.getRequestFactory().get(path);
 
         MantaObjectInputStream stream = httpHelper.httpRequestAsInputStream(get,
                 requestHeaders);
@@ -582,8 +562,7 @@ public class MantaClient implements AutoCloseable {
         Validate.notBlank(rawPath, "Path must not be blank");
         String path = formatPath(rawPath);
 
-        return new MantaSeekableByteChannel(path, position,
-                this.connectionFactory, this.httpHelper);
+        return new MantaSeekableByteChannel(path, position, httpHelper);
     }
 
     /**
@@ -602,8 +581,7 @@ public class MantaClient implements AutoCloseable {
         Validate.notBlank(rawPath, "Path must not be blank");
         String path = formatPath(rawPath);
 
-        return new MantaSeekableByteChannel(path, this.connectionFactory,
-                this.httpHelper);
+        return new MantaSeekableByteChannel(path, httpHelper);
     }
 
     /**
@@ -1490,7 +1468,7 @@ public class MantaClient implements AutoCloseable {
 
         LOG.debug("PUT    {} [directory]", path);
 
-        final HttpPut put = connectionFactory.put(path);
+        final HttpPut put = httpHelper.getRequestFactory().put(path);
 
         final MantaHttpHeaders headers;
 
@@ -1581,7 +1559,7 @@ public class MantaClient implements AutoCloseable {
         final String objectPath = formatPath(rawObjectPath);
 
         LOG.debug("PUT    {} -> {} [snaplink]", objectPath, linkPath);
-        final HttpPut put = connectionFactory.put(linkPath);
+        final HttpPut put = httpHelper.getRequestFactory().put(linkPath);
 
         if (headers != null) {
             put.setHeaders(headers.asApacheHttpHeaders());
@@ -1752,7 +1730,7 @@ public class MantaClient implements AutoCloseable {
         HttpEntity entity = new ExposedByteArrayEntity(json,
                 ContentType.APPLICATION_JSON);
 
-        HttpPost post = connectionFactory.post(path);
+        HttpPost post = httpHelper.getRequestFactory().post(path);
         post.setEntity(entity);
 
         Function<CloseableHttpResponse, UUID> jobIdFunction = response -> {
@@ -1854,7 +1832,7 @@ public class MantaClient implements AutoCloseable {
 
         String path = String.format("%s/jobs/%s/live/in", home, jobId);
 
-        HttpPost post = connectionFactory.post(path);
+        HttpPost post = httpHelper.getRequestFactory().post(path);
         post.setHeader(HttpHeaders.CONTENT_ENCODING, "chunked");
         post.setEntity(entity);
 
@@ -1875,7 +1853,7 @@ public class MantaClient implements AutoCloseable {
         Validate.notNull(jobId, "Manta job id must not be null");
         String path = String.format("%s/jobs/%s/live/in", home, jobId);
 
-        HttpGet get = connectionFactory.get(path);
+        HttpGet get = httpHelper.getRequestFactory().get(path);
         HttpResponse response = httpHelper.executeRequest(get,
                 "GET    {} response [{}] {} ");
         return responseAsStream(response);
@@ -1939,8 +1917,8 @@ public class MantaClient implements AutoCloseable {
         final String livePath = String.format("%s/jobs/%s/live/status",
                 home, jobId);
 
-        final CloseableHttpClient client = connectionContext.getHttpClient();
-        final HttpUriRequest initialRequest = connectionFactory.get(livePath);
+        final CloseableHttpClient client = httpHelper.getConnectionContext().getHttpClient();
+        final HttpUriRequest initialRequest = httpHelper.getRequestFactory().get(livePath);
         MantaJob job;
         HttpEntity entity;
 
@@ -1956,7 +1934,7 @@ public class MantaClient implements AutoCloseable {
                 final String archivePath = String.format("%s/jobs/%s/job.json",
                         home, jobId);
 
-                final HttpUriRequest archiveRequest = connectionFactory.get(archivePath);
+                final HttpUriRequest archiveRequest = httpHelper.getRequestFactory().get(archivePath);
 
                 // We close the request that was previously opened, because it
                 // didn't have what we need.
@@ -2226,7 +2204,7 @@ public class MantaClient implements AutoCloseable {
         }
 
         final String path = formatPath(String.format("%s/jobs", home));
-        final HttpGet get = connectionFactory.get(path, params);
+        final HttpGet get = httpHelper.getRequestFactory().get(path, params);
 
         final HttpResponse response = httpHelper.executeRequest(get,
                 "GET    {} response [{}] {} ");
@@ -2273,7 +2251,7 @@ public class MantaClient implements AutoCloseable {
         Validate.notNull(jobId, "Job id must not be null");
         String path = String.format("%s/jobs/%s/live/out", home, jobId);
 
-        HttpGet get = connectionFactory.get(path);
+        HttpGet get = httpHelper.getRequestFactory().get(path);
         HttpResponse response = httpHelper.executeRequest(get,
                 "GET    {} response [{}] {} ");
         return responseAsStream(response);
@@ -2356,7 +2334,7 @@ public class MantaClient implements AutoCloseable {
 
         String path = String.format("%s/jobs/%s/live/fail", home, jobId);
 
-        final HttpGet get = connectionFactory.get(path);
+        final HttpGet get = httpHelper.getRequestFactory().get(path);
         final HttpResponse response = httpHelper.executeRequest(get,
                 "GET    {} response [{}] {} ");
 
@@ -2379,7 +2357,7 @@ public class MantaClient implements AutoCloseable {
 
         final String path = String.format("%s/jobs/%s/live/err", home, jobId);
 
-        final HttpGet get = connectionFactory.get(path);
+        final HttpGet get = httpHelper.getRequestFactory().get(path);
         final HttpResponse response = httpHelper.executeRequest(get,
                 "GET    {} response [{}] {} ");
         final ObjectMapper mapper = MantaObjectMapper.INSTANCE;
@@ -2483,12 +2461,6 @@ public class MantaClient implements AutoCloseable {
             /* Do nothing, but we won't capture the interrupted exception
              * because even if we are interrupted, we want to close all open
              * resources. */
-        } catch (Exception e) {
-            exceptions.add(e);
-        }
-
-        try {
-            connectionFactory.close();
         } catch (Exception e) {
             exceptions.add(e);
         }