+++ java-manta-client-unshaded/src/main/java/com/joyent/manta/util/MantaVersion.java
@@ -22,7 +22,7 @@ public final class MantaVersion {
     /**
      * Release version of the SDK.
      */
-    public static final String VERSION = "3.0.0-SNAPSHOT";
+    public static final String VERSION = "3.1.7-SNAPSHOT";
 
     /**
      * Minimum version of client-side encryption supported.